Add FileOptions::mmap_size for mmap larger than file size
mmap_size allow us create a huge mmap for reserving a virtual address
range, later we read the really written data, this is very useful for
memtable_as_log_index. Previously we truncate the file to a huge size
for such purpose, but the side effect is that external process can not
tailing the file.

By using this feature, all requirements are satified!

Linux mmap support this feature, windows will be considered
later.