Use vector io log::Writer::AddRecordv() to avoid memcpy
WriteBatchInternal::Append expand and copy batches in WriteGroup,
this is not needed.

We gather the batches's memory and use `AddRecordv()` to calling
`writev`, this avoiding memory allocation and memcpy.

This is much faster than upstream rocksdb in multi thread write,
the performance degration of old code is not obvious because rocksdb
is slow. With our `memtable_as_log_index`, write is very fast, so the
bottle neck of old code emerges.