TxnDB: zero copy prepared WriteBatch in commit
In txn commit, rocksdb append prepared (WBWI)WriteBatch to commit
time WriteBatch. To skip writing prepared WriteBatch again, rocksdb
add `WalTerminationPoint` to WriteBatch. This not only need extra
memcpy but also increase complicity.

In this commit, we link the prepared WriteBatch to commit WriteBatch
and remove `WalTerminationPoint`. This removed the memcpy, and
simplified the code.