java: RocksIterator old api support zero copy - 2
If env ROCKSDB_FORCE_DIRECT_BUFFER_ZERO_COPY is true
(default is false if not defined),
even the ByteBuffer is created by `ByteBuffer.allocateDirect(cap)`,
zero copy will also be applied, the allocated memory will be
free'ed by calling DirectByteBuffer.cleaner.clean(), and the
cleaner is set to null.

This is not recommonded, but works for most existing code,
thus minimize the migration efforts.

This need java cmd option:
   --add-opens java.base/jdk.internal.ref=ALL-UNNAMED